+++ build.rs
@@ -78,6 +78,7 @@ fn parse(contents: String, file_name: String) -> ellie_parser::parser::ParserRes
             loops: true,
             conditions: true,
             classes: true,
+            enums: true,
             dynamics: true,
             global_variables: true,
             getters: true,